All goserver ported to use Redis backend
All of goserver now uses a database wrapper struct that has to implement several functions. I'm still in the process of tightening this up so it's less brittle and well structured but this is the first version that works with all existing goquery functionality (though some hidden data such as query status from the remote host have been lost).

Redis as the only backend is not a super good idea though it is durable if configured as such (currently it's because these docker setups are supposed to be ephemeral anyway) but this is how one could use Redis as a full cache for an osquery distributed server. All enrollment, checkins, query schedules, and query results are stored in Redis and fetched when needed. There are also no scans so all look ups should be quick and Redis shouldn't spend anytime returning large amounts of unneeded data and goserver doesn't need to filter anything. This does mean there is a slight bit of data duplication but that might be able to be ironed out in the future. I've documented the "schema" I've created below.

Top level Redis key scan of goquery testing infra redis server after running some queries on one host:
```
127.0.0.1:6379> scan 0
1) "0"
2) 1) "nodekey:OA5xUCsRe3qeBv6HgjCdRQGHCJUtUmQN"
   2) "queries:nodekey:OA5xUCsRe3qeBv6HgjCdRQGHCJUtUmQN"
   3) "queries:complete"
   4) "nodekey:ItGT7FoCy5vM5XJTXN43rnYgrzlg4r3b"
   5) "uuid:06e0d506-1c83-4396-826b-bbe7022aa4eb"
   6) "uuid:cfa92f2c-3750-442c-a604-472aa366b579"
```

**Enrollment:**
When a host checks in and enrolls, two keys are created:
    uuid:<uuid> -> <nodekey>
    nodekey:<nodekey> -> HMSET of host info

This allows quick look up by UUID or nodekey without duplicating the host info.

**Query Scheduling**
When a query is scheduled against a host we need to do a few things to avoid scans:
    0) Add the query name into the full history query map for the host: queries:nodekey:<nodekey> with the key as <queryName> and the value the query that needs to be run
    1) Add the query name into the global pending query map: queries:pending with key <queryName> and value of query to be run
    2) Add the query name into the nodekey pending query map: queries:pending:nodekey:<nodekey> with the same key and value as above

The query is now scheduled to be run

**Query Pickup**
We did extra work in query scheduling so query pickup is easy. This was done because scheduling happens much more infrequently than pick up so we want to bias that to be fast. Query pickup is basically returning the HMSet: queries:pending:nodekey:<nodekey>

**Result Storage**
Result storage also has a few steps because the of the optimization for query pickup means it needs to clean up pending status. This also means that storage currently needs to know the nodekey of the host that sent the results to remove it from the pending set but that may be fixed in the future by adding a top level lookup HMSet.

When storage happens there are three database operations:
    0) Store the results under the map: queries:complete with key <queryName>
    1) Delete <queryName> from the global pending map: queries:pending
    2) Delete <queryName> from the nodekey pending map: queries:pending:nodekey

One other thing to note is that this diff reduces the query name from 64 characters to 32. Given our character set for query names is currently 62 characters long (26 + 26 + 10) this gives us an effective 192 bits of entropy (down from 384 with 64 characters). This should be more than enough and makes the UX in goquery nicer when dealing with tab complete.